+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This is a manually written changelog, and only tracks front-facing changes since
 - `download`: try to add entries from the recovery to the archive after move
 - `download`: add option `--extra-ytdl-args` to provide extra youtube-dl arguments
 - `download`: does not immediately error anymore when a "ERROR:" line in encountered (like a private video in a playlist)
+- `download`: reflect skipped and errored media in the count
 - ffmpeg: fix error on invalid utf8 sequence
 - add backtraces to errors that do not panic (only when `RUST_BACKTRACE=true` is set)
 - change Termination requests to not be based on time anymore